+++ pybv/tests/test_bv_writer.py
@@ -11,6 +11,7 @@
 # License: BSD (3-clause)
 
 from datetime import datetime, timezone
+import re
 
 import mne
 import numpy as np
@@ -34,6 +35,10 @@ event_times = np.arange(1, 5)
 events = np.column_stack([event_times * sfreq, [1, 1, 2, 2]])
 # scale random data to reasonable EEG signal magnitude in V
 data = rng.randn(n_chans, n_times) * 10 * 1e-6
+
+# add reference channel
+ref_ch_name = ch_names[-1]
+data[-1, :] = 0.
 
 
 @pytest.mark.parametrize(
@@ -86,6 +91,10 @@ def test_bv_writer_inputs(tmpdir):
         write_brainvision(data=data, sfreq=sfreq, ch_names=ch_names,
                           fname_base=fname, folder_out=tmpdir,
                           resolution=np.arange(n_chans-1))
+    with pytest.raises(ValueError, match='reference channel.*not found'):
+        write_brainvision(data=data, sfreq=sfreq, ch_names=ch_names,
+                          ref_ch_name='foobar', fname_base=fname,
+                          folder_out=tmpdir)
 
 
 def test_bv_bad_format(tmpdir):
@@ -95,13 +104,16 @@ def test_bv_bad_format(tmpdir):
     eeg_fname = tmpdir / fname + ".eeg"
 
     with pytest.raises(ValueError, match='Orientation bad not supported'):
-        _write_vhdr_file(vhdr_fname, vmrk_fname, eeg_fname, data=data,
-                         sfreq=sfreq, ch_names=ch_names, orientation='bad',
+        _write_vhdr_file(vhdr_fname=vhdr_fname, vmrk_fname=vmrk_fname,
+                         eeg_fname=eeg_fname, data=data,
+                         sfreq=sfreq, ch_names=ch_names, ref_ch_name=None,
+                         orientation='bad',
                          format="binary_float32", resolution=1e-6,
                          units=["V"] * n_chans)
     with pytest.raises(ValueError, match='Data format bad not supported'):
-        _write_vhdr_file(vhdr_fname, vmrk_fname, eeg_fname, data=data,
-                         sfreq=sfreq, ch_names=ch_names,
+        _write_vhdr_file(vhdr_fname=vhdr_fname, vmrk_fname=vmrk_fname,
+                         eeg_fname=eeg_fname, data=data,
+                         sfreq=sfreq, ch_names=ch_names, ref_ch_name=None,
                          orientation='multiplexed', format="bad",
                          resolution=1e-6,
                          units=["V"] * n_chans)
@@ -147,24 +159,29 @@ def test_comma_in_ch_name(tmpdir, ch_names_tricky):
     assert_allclose(data, raw_written._data)  # data round-trip
 
 
-@pytest.mark.parametrize("meas_date",
-                         [('20000101120000000000'),
-                          (datetime(2000, 1, 1, 12, 0, 0, 0))])
-def test_write_read_cycle(tmpdir, meas_date):
+@pytest.mark.parametrize(
+    "meas_date, ref_ch_name",
+    [
+        ('20000101120000000000', ref_ch_name),
+        (datetime(2000, 1, 1, 12, 0, 0, 0), None)
+    ]
+)
+def test_write_read_cycle(tmpdir, meas_date, ref_ch_name):
     """Test that a write/read cycle produces identical data."""
     # First fail writing due to wrong unit
     unsupported_unit = "rV"
     with pytest.warns(UserWarning, match='Encountered unsupported '
                                          'non-voltage unit'):
         write_brainvision(data=data, sfreq=sfreq, ch_names=ch_names,
-                          fname_base=fname, folder_out=tmpdir,
-                          unit=unsupported_unit)
+                          ref_ch_name=ref_ch_name, fname_base=fname,
+                          folder_out=tmpdir, unit=unsupported_unit)
 
     # write and read data to BV format
     # ensure that greek small letter mu gets converted to micro sign
     with pytest.warns(UserWarning, match="Encountered small Greek letter mu"):
         write_brainvision(data=data, sfreq=sfreq, ch_names=ch_names,
-                          fname_base=fname, folder_out=tmpdir, events=events,
+                          ref_ch_name=ref_ch_name, fname_base=fname,
+                          folder_out=tmpdir, events=events,
                           resolution=np.power(10., -np.arange(10)),
                           unit='μV', meas_date=meas_date)
     vhdr_fname = tmpdir / fname + '.vhdr'
@@ -325,3 +342,40 @@ def test_write_unsupported_units(tmpdir):
     assert len(set(orig_units)) == 2
     assert all([orig_units[idx] == unit for idx in range(n_chans - 1)])
     assert orig_units[-1] == '°C'
+
+
+def test_ref_ch(tmpdir):
+    """Test reference channel writing."""
+    # these are the default values
+    resolution = '0.1'
+    unit = 'µV'
+
+    vhdr_fname = tmpdir / fname + '.vhdr'
+
+    # Passing a ref_ch_name should write it to the VHDR file
+    write_brainvision(data=data, sfreq=sfreq, ch_names=ch_names,
+                      ref_ch_name=ref_ch_name, fname_base=fname,
+                      folder_out=tmpdir)
+
+    vhdr = vhdr_fname.read_text(encoding='utf-8')
+    regexp = f'Ch.*=ch.*,{ref_ch_name},{resolution},{unit}'
+    matches = re.findall(pattern=regexp, string=vhdr)
+    assert len(matches) == len(ch_names)
+
+    # Not passing a ref_ch_name should omit the reference channel field
+    write_brainvision(data=data, sfreq=sfreq, ch_names=ch_names,
+                      fname_base=fname, folder_out=tmpdir)
+
+    vhdr = vhdr_fname.read_text(encoding='utf-8')
+    regexp = f'Ch.*=ch.*,,{resolution},{unit}'
+    matches = re.findall(pattern=regexp, string=vhdr)
+    assert len(matches) == len(ch_names)
+
+    # Passing data that's not all-zero for a reference channel should raise
+    # an exception
+    data_ = data.copy()
+    data_[ch_names.index(ref_ch_name), :] = 5
+    with pytest.raises(ValueError, match='reference channel.*not.*zero'):
+        write_brainvision(data=data_, sfreq=sfreq, ch_names=ch_names,
+                          ref_ch_name=ref_ch_name, fname_base=fname,
+                          folder_out=tmpdir)