+++ docs/changelog.rst
@@ -29,7 +29,10 @@ Here we list a changelog of pybv.
 Current (unreleased)
 ====================
 
-- nothing so far
+Changelog
+~~~~~~~~~
+
+- :func:`pybv.write_brainvision` gained a new parameter, ``ref_ch_name``, to specify the reference channel used during recording, by `Richard Höchenberger`_ (:gh:`75`)
 
 0.5.0 (2021-01-03)
 ==================